requirements.txt
APScheduler==3.10.4
async-timeout==4.0.2
backports.zoneinfo==0.2.1
bcrypt==4.0.1
blinker==1.6.2
click==8.1.6
colorama==0.4.6
et-xmlfile==1.1.0
Flask==2.3.2
Flask-APScheduler==1.12.4
Flask-Cors==4.0.0
flask-sqlalchemy==3.0.5
greenlet==2.0.2
importlib-metadata==6.8.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.4
openpyxl==3.1.2
pandas==2.0.3
Pillow==10.0.0
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
six==1.16.0
SQLAlchemy==2.0.19
treelib==1.6.4
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
Werkzeug==2.3.6
zipp==3.16.2
